Create Top_Words
You can use the 'phrases.json' endpoint to find top words or phrases for a certain type of entity. The entity type can be a state, legislator, date, or month. When making requests to this endpoint, you need to specify an 'entity_type' and an 'entity_value'.

Because this particular endpoint returns the most spoken phrases, you can only choose one type of entity (state, legislator, date or month) to get the top phrases for. This is why the parameter names are a little different for this endpoint. Due to the sheer volume of text in the Congressional Record, we can only support getting the top phrases by one entity_type at this time.

Each 'entity_type' has its own format for 'entity_value':

state: 'AL' (two letter abbreviation)
legislator: 'L000551' (a Bioguide ID for the legislator)
month: '201211' (4 digit year, 2 digit month)
date: '2012-11-30'
Sunlight Note (optional) :
Bioguide ids are maintained by the Congress. Each legislator throughout history has a unique one. You can look up any legislator in the Bioguide and infer their Bioguide ID from the URL. 
For instance, if you look at the URL for Thomas Jefferson: 
http://bioguide.congress.gov/scripts/biodisplay.pl?index=J000069

the bioguide id is 'J000069'.

Bioguide ids appear in a lot of our APIs because they are a great way of tracking legislators uniquely.