Create user_search_engine
"So from what I've learned after messing around with this for a while is that in order to put the field you want into a variable, you need to narrow it down. data is a dictionary, which contains num_found and results. But results is a list of many different occurrences, based on your per_page value. These occurrences are actually all dictionaries. Following? So within the data dict, there is a results list, which contains occurrence dicts based on what your per_page number is. You can see in my code, I did index 0 within results because I want the soonest occurrence (since I sorted by date desc, the most recent is at index 0) . The keys within these occurrence dictionaries are standard. You can see that in order to get the date, speaker, and speech values, I drew the values from the keys, date,speaker_raw, and speaking respectively."

"Another note: It's better to use speaker_raw, because it always has a value, whereas speaker_first and speaker_last often have a value of None."